+++ src/chatbot/config.js
@@ -1,0 +1,18 @@
+import { createChatBotMessage } from "react-chatbot-kit";
+
+const botName = 'A.S.';
+
+const config = {
+  initialMessages: [createChatBotMessage(`Hi! I'm ${botName}`)],
+  botName: botName,
+  customStyles: {
+    botMessageBox: {
+      backgroundColor: '#376B7E',
+    },
+    chatButton: {
+      backgroundColor: '#5ccc9d',
+    },
+  },
+};
+
+export default config
