--- src/config.js
@@ -1,7 +1,0 @@
-import { createChatBotMessage } from "react-chatbot-kit";
-
-const config = {
-  initialMessages: [createChatBotMessage(`Hello world`)]
-}
-
-export default config
